pyproject.toml
[tool.poetry]
name = "kernel-telegram-bot"
version = "2025.15.15"
description = "Kernel for a telegram bot based on python-telegram-bot"
authors = ["LoboGuardian <loboguardian.dev@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["webhooks"], version = "^22.0"}
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
